Rename _nested_seq.py to _nested_iter.py
