+++ skbase/utils/_nested_iter.py
@@ -8,6 +8,23 @@ __author__: List[str] = ["RNKuhns", "fkiraly"]
 __all__: List[str] = ["flatten", "is_flat", "unflat_len", "unflatten"]
 
 
+def _remove_single(x):
+    """Remove tuple wrapping from singleton.
+
+    Parameters
+    ----------
+    x : tuple
+
+    Returns
+    -------
+    x[0] if x is a singleton, otherwise x
+    """
+    if len(x) == 1:
+        return x[0]
+    else:
+        return x
+
+
 def flatten(obj):
     """Flatten nested list/tuple structure.
